requirements.txt
# Core data stack
pandas>=2.0
numpy>=1.24
scipy>=1.11

# Geospatial
geopandas>=0.14
rasterio>=1.3
rioxarray>=0.15
shapely>=2.0
pyproj>=3.6

# Satellite & climate data access
pystac>=1.9
pystac_client>=0.7
planetary_computer>=1.0
odc-stac>=0.3
xarray[complete]>=2024.0
netCDF4>=1.6
zarr>=2.16
dask>=2024.0
adlfs>=2024.0
numcodecs>=0.12

# Machine learning
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
optuna>=3.4

# Model interpretation
shap>=0.43

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Utilities
tqdm>=4.66
pyyaml>=6.0
joblib>=1.3
